Fixed critical bug with getting a friend number given an ID, it was making deleting friends impossible. Also cleaned up some various parts in the code.

+++ Toxicity/TXCAppDelegate.h
@@ -57,7 +57,7 @@ int friendNumForID(NSString *theKey);
 - (void)acceptFriendRequests:(NSArray *)theKeysToAccept;
 - (void)acceptGroupInvites:(NSArray *)theKeysToAccept;
 - (int)deleteFriend:(NSString*)friendKey;
-- (int)deleteGroupchat:(int)theGroupNumber;
+- (int)deleteGroupchat:(NSInteger)theGroupNumber;
 
 - (void)configureNavigationControllerDesign:(UINavigationController *)navController;
